+++ src/components/ElementBar/DuplicateActiveElementButton/DuplicateActiveElementButton.test.tsx
@@ -15,6 +15,7 @@
  */
 
 import { Element } from '../../../state';
+import { calculateBoundingRectForElements } from '../../../state/crdt/documents/elements';
 import { duplicate } from './DuplicateActiveElementButton';
 
 describe('duplicate', () => {
@@ -30,7 +31,9 @@ describe('duplicate', () => {
       strokeColor: '#ff0000',
     };
 
-    expect(duplicate(element, 40)).toEqual({
+    expect(
+      duplicate(element, 40, calculateBoundingRectForElements([element])),
+    ).toEqual({
       ...element,
       // 10 is x, 80 is the width and 40 is the grid cell size
       position: { x: 10 + 80 + 40, y: 20 },
@@ -48,8 +51,57 @@ describe('duplicate', () => {
       text: '',
     };
 
-    expect(duplicate(element, 40)).toEqual({
+    expect(
+      duplicate(element, 40, calculateBoundingRectForElements([element])),
+    ).toEqual({
       ...element,
+      // 10 is x, 80 is the width and 40 is the grid cell size
+      position: { x: 10 + 80 + 40, y: 20 },
+    });
+  });
+
+  it('should duplicate multiple elements', () => {
+    const element0: Element = {
+      type: 'path',
+      kind: 'line',
+      position: { x: 10, y: 20 },
+      points: [
+        { x: 0, y: 0 },
+        { x: 80, y: 70 },
+      ],
+      strokeColor: '#ff0000',
+    };
+
+    const element1: Element = {
+      type: 'shape',
+      kind: 'rectangle',
+      position: { x: 10, y: 20 },
+      width: 80,
+      height: 70,
+      fillColor: '#000',
+      text: '',
+    };
+
+    expect(
+      duplicate(
+        element0,
+        40,
+        calculateBoundingRectForElements([element0, element1]),
+      ),
+    ).toEqual({
+      ...element0,
+      // 10 is x, 80 is the width and 40 is the grid cell size
+      position: { x: 10 + 80 + 40, y: 20 },
+    });
+
+    expect(
+      duplicate(
+        element1,
+        40,
+        calculateBoundingRectForElements([element0, element1]),
+      ),
+    ).toEqual({
+      ...element1,
       // 10 is x, 80 is the width and 40 is the grid cell size
       position: { x: 10 + 80 + 40, y: 20 },
     });
@@ -67,7 +119,9 @@ describe('duplicate', () => {
       strokeColor: '#ff0000',
     };
 
-    expect(duplicate(element, 40)).toEqual({
+    expect(
+      duplicate(element, 40, calculateBoundingRectForElements([element])),
+    ).toEqual({
       ...element,
       // 80 is the width
       position: { x: 1920 - 80, y: 20 },
@@ -85,8 +139,57 @@ describe('duplicate', () => {
       text: '',
     };
 
-    expect(duplicate(element, 40)).toEqual({
+    expect(
+      duplicate(element, 40, calculateBoundingRectForElements([element])),
+    ).toEqual({
       ...element,
+      // 80 is the width
+      position: { x: 1920 - 80, y: 20 },
+    });
+  });
+
+  it('should stop at the right edge when duplicating multiple elements', () => {
+    const element0: Element = {
+      type: 'path',
+      kind: 'line',
+      position: { x: 1920 - 120, y: 20 },
+      points: [
+        { x: 0, y: 0 },
+        { x: 80, y: 70 },
+      ],
+      strokeColor: '#ff0000',
+    };
+
+    const element1: Element = {
+      type: 'shape',
+      kind: 'rectangle',
+      position: { x: 1920 - 120, y: 20 },
+      width: 80,
+      height: 70,
+      fillColor: '#000',
+      text: '',
+    };
+
+    expect(
+      duplicate(
+        element0,
+        40,
+        calculateBoundingRectForElements([element0, element1]),
+      ),
+    ).toEqual({
+      ...element0,
+      // 80 is the width
+      position: { x: 1920 - 80, y: 20 },
+    });
+
+    expect(
+      duplicate(
+        element1,
+        40,
+        calculateBoundingRectForElements([element0, element1]),
+      ),
+    ).toEqual({
+      ...element1,
       // 80 is the width
       position: { x: 1920 - 80, y: 20 },
     });
@@ -104,7 +207,9 @@ describe('duplicate', () => {
       strokeColor: '#ff0000',
     };
 
-    expect(duplicate(element, 40)).toEqual({
+    expect(
+      duplicate(element, 40, calculateBoundingRectForElements([element])),
+    ).toEqual({
       ...element,
       // 80 is the width
       position: { x: 1920 - 80, y: 20 },
@@ -122,8 +227,57 @@ describe('duplicate', () => {
       text: '',
     };
 
-    expect(duplicate(element, 40)).toEqual({
+    expect(
+      duplicate(element, 40, calculateBoundingRectForElements([element])),
+    ).toEqual({
       ...element,
+      // 80 is the width
+      position: { x: 1920 - 80, y: 20 },
+    });
+  });
+
+  it('should stay put if already at the right edge when duplicating multiple elements', () => {
+    const element0: Element = {
+      type: 'path',
+      kind: 'line',
+      position: { x: 1920 - 80, y: 20 },
+      points: [
+        { x: 0, y: 0 },
+        { x: 80, y: 70 },
+      ],
+      strokeColor: '#ff0000',
+    };
+
+    const element1: Element = {
+      type: 'shape',
+      kind: 'rectangle',
+      position: { x: 1920 - 80, y: 20 },
+      width: 80,
+      height: 70,
+      fillColor: '#000',
+      text: '',
+    };
+
+    expect(
+      duplicate(
+        element0,
+        40,
+        calculateBoundingRectForElements([element0, element1]),
+      ),
+    ).toEqual({
+      ...element0,
+      // 80 is the width
+      position: { x: 1920 - 80, y: 20 },
+    });
+
+    expect(
+      duplicate(
+        element1,
+        40,
+        calculateBoundingRectForElements([element0, element1]),
+      ),
+    ).toEqual({
+      ...element1,
       // 80 is the width
       position: { x: 1920 - 80, y: 20 },
     });