+++ src/state/crdt/documents/index.ts
@@ -26,6 +26,7 @@ export type {
 } from './elements';
 export {
   generateAddElement,
+  generateAddElements,
   generateAddSlide,
   generateDuplicateSlide,
   generateLockSlide,